[refactor] #169 디렉토리 구조 수정
