--- src/main/java/org/farmsystem/homepage/domain/minigame/entity/PlantStatus.java
@@ -1,5 +1,0 @@
-package org.farmsystem.homepage.domain.minigame.entity;
-
-public enum PlantStatus {
-    PLANTED, FLOWERED, NOTHING
-}
