--- src/main/java/org/farmsystem/homepage/domain/minigame/entity/PlantType.java
@@ -1,5 +1,0 @@
-package org.farmsystem.homepage.domain.minigame.entity;
-
-public enum PlantType {
-    TREE, FLOWER
-}
