--- src/main/java/org/farmsystem/homepage/domain/minigame/entity/Rotation.java
@@ -1,5 +1,0 @@
-package org.farmsystem.homepage.domain.minigame.entity;
-
-public enum Rotation {
-    R0, R90, R180, R270
-}
