+++ src/main/java/org/farmsystem/homepage/domain/minigame/farm/entity/FarmplotTile.java
@@ -1,10 +1,11 @@
-package org.farmsystem.homepage.domain.minigame.entity;
+package org.farmsystem.homepage.domain.minigame.farm.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.farmsystem.homepage.domain.minigame.player.entity.Player;
 
 @Entity
 @Table(name = "farmplot_tile")
