+++ src/main/java/org/farmsystem/homepage/domain/minigame/farm/entity/PlantStatus.java
@@ -1,0 +1,5 @@
+package org.farmsystem.homepage.domain.minigame.farm.entity;
+
+public enum PlantStatus {
+    PLANTED, FLOWERED, NOTHING
+}
