+++ src/main/java/org/farmsystem/homepage/domain/minigame/farm/entity/PlantType.java
@@ -1,0 +1,5 @@
+package org.farmsystem.homepage.domain.minigame.farm.entity;
+
+public enum PlantType {
+    TREE, FLOWER
+}
