+++ src/main/java/org/farmsystem/homepage/domain/minigame/farm/entity/PlantedPlant.java
@@ -1,7 +1,9 @@
-package org.farmsystem.homepage.domain.minigame.entity;
+package org.farmsystem.homepage.domain.minigame.farm.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.farmsystem.homepage.domain.minigame.player.entity.Player;
+
 import java.time.LocalDateTime;
 
 @Entity
