+++ src/main/java/org/farmsystem/homepage/domain/minigame/garden/entity/Rotation.java
@@ -1,0 +1,5 @@
+package org.farmsystem.homepage.domain.minigame.garden.entity;
+
+public enum Rotation {
+    R0, R90, R180, R270
+}
