+++ src/main/java/org/farmsystem/homepage/domain/minigame/inventory/entity/ObjectInventory.java
@@ -1,7 +1,8 @@
-package org.farmsystem.homepage.domain.minigame.entity;
+package org.farmsystem.homepage.domain.minigame.inventory.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
+import org.farmsystem.homepage.domain.minigame.player.entity.Player;
 
 @Entity
 @Table(name = "object_inventory")
