+++ src/main/java/org/farmsystem/homepage/domain/minigame/player/entity/DailyGame.java
@@ -1,4 +1,4 @@
-package org.farmsystem.homepage.domain.minigame.entity;
+package org.farmsystem.homepage.domain.minigame.player.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
